Car-Orientation-Detection/requirements.txt
numpy
paddlepaddle-gpu==2.3.2
opencv-python==3.4.18
scikit-image
imageio